Added subscriber that subscribes to the 'aruco_detections' topic and stores detected aruco tags.
I added subscriber that subscribes to the 'aruco_detections' topic and stores detected aruco tags. I figured that a good way to implement this would be with a queue implemented as a list in python. Since this is an inline script (not a class), I simply made both the list and the subscribers global variables. The list can be accessed within a given scope by defining it locally as the global variable using the global keyword. The best way to finish this implementation will likely be to have a "act according to tags" method of some sore that is simply constantly run that will retrieve the first not-yet-handled aruco tag from the global list and then call the correct movement method to execute the appropriate action. After that it will remove it from the list and move on to the next tag in the list, if there are any.