Added in Publishing to movement_test Capability
I added in the functionality of publishing to the '/aruco_detections' topic which is subscribed to by movement_test.py. This connects the two ROS nodes and allows them to communicate. Currently on the id of the aruco tag is published to the topic. If we want, the entire aruco tag's MarkerArray could be published instead, and that would be a relatively simple change of just changing the message type in the aruco_publisher initialization and then publishing the data returned from cv2.aruco.detectMarkers(gray, aruco dict, parameters=parameters). However, it may be more beneficial to split the detections into individual aruco tags and that would require a custom message type containing the corners and ids for each tag. This also isn't terrible but would require a bit more work. Let me know if you want me to work on that.

Side note:
writing this made me realize that we probably want a set-like implementation of the aruco tag id storing datatype in movement_test.py because if an aruco tag is detected twice before the robot handles it, it may cause it to execute the corresponding command several times even after the tag leaves view... I don't know though and will hold off on making this change until I talk to @TawhidAhmad